+++ src/mirai-translator.ts
@@ -2,9 +2,34 @@ import * as request from 'request-promise';
 import * as vscode from 'vscode';
 const cheerio = require('cheerio');
 
+let proxyIsEnable: boolean = true;
+
+export function enableProxy() {
+    proxyIsEnable = true;
+}
+
+export function disableProxy() {
+    proxyIsEnable = false;
+}
+
 export enum LanguageEnum {
     JAPANESE = 'ja',
     ENGLISH = 'en',
+}
+
+/**
+ * Returns proxy setting that a user configured
+ *
+ * @returns null if proxy is not configured
+ */
+function getProxy(): string | null {
+    const config = vscode.workspace.getConfiguration('miraiTranslator.proxy');
+
+    if (!config.get("host") || !proxyIsEnable) {
+        return null;
+    }
+
+    return `http://${config.get("username")}:${config.get("password")}@${config.get("host")}:${config.get("port")}`;
 }
 
 export function translate(textToBeTranslated: string | undefined, source: LanguageEnum,
@@ -24,7 +49,10 @@ export function translate(textToBeTranslated: string | undefined, source: Langua
         method: 'GET'
     };
 
-    const req = request.defaults({ jar: true });
+    const req = request.defaults({
+        jar: true,
+        proxy: getProxy(),
+    });
 
     req(requestOptionsToGetCookie)
         .then(function (body: any) {